ts props
