+++ src/client/rsg-components/Props/util.ts
@@ -9,6 +9,8 @@ export function unquote(string?: string): string | undefined {
 
 export interface PropDescriptorWithFlow extends PropDescriptor {
 	flowType?: FlowTypeDescriptor;
+	name: string;
+	tags?: Record<string, any>;
 }
 
 /**
